Use a ZIF footprint for U5 (EEPROM)
ZIF sockets have feet that are slightly too wide to fit in a regular DIP
footprint's drill holes. This caused me issues in the fist spin of
FUNCOID1. Replace it with a ZIF footprint as was originally envisaged.

I had to upgrade to KiCad 5 to make use of these new footprints. I made
a custom derivative where the ZIF lever is facing the other way, but
preserving the orientation of the pads.

Some components had to be shifted around a little to avoid overlapping
courtyards.