Add a MAX232 for proper RS-232 signalling
Use a MAX232 to drive the signal on the DB9 connector, such that it is
RS-232 compliant and will be understood by my VT510 terminal. Route the
old 5V CMOS UART to a new 6-pin header for use with an FTDI.

This change also moves the DB9 connector to the same edge as the barrel
jack, which should make designing a box a little easier.

Finally, do some general re-routing on the PCB layout.